添加设置页面、使用第三方alertdialog
github上找了个很炫酷的alertdialog
设置还是比较丑
明天做主页 用图表显示最近情况（？）

+++ app/src/main/java/com/crossbow/app/x_timer/Utils/FileUtils.java
@@ -155,9 +155,20 @@ public class FileUtils {
         File root = new File("/data/data/com.crossbow.app.x_timer/files");
         File[] files = root.listFiles();
         for (File file : files) {
-            file.delete();
+            if (file.getName().startsWith("flag_")) file.delete();
         }
 
+        return true;
+    }
+
+    // 删除所监听列表
+    public boolean deleteAppList() {
+        File file = new File("/data/data/com.crossbow.app.x_timer/files/appList");
+        if (!file.exists()) {
+            return false;
+        }
+
+        file.delete();
         return true;
     }
 }
